Use Gradle as the build tool, replacing Moxie.
Standard Gradle tasks are used in the build.gradle script to compile and
package Gitblit artefacts. Also, custom Gradle tasks are added to the
standard buildSrc folder to achieve complex operations which are unique
to the Gitblit project, and thus are unavailable in standard Gradle or
any plugins.

Gradle itself need JDK8 or later versions to run, and by default the
same JDK is also used by Gradle to compile the Gitblit project's code.
But due to current targetCompatibility = '1.7' setting, the resulting
build artefacts are still compatible with the Java7 runtime.

If the JDK7 compiler is required, some settings to use the JDK7
installation are already in the build.gradle script, but are currently
commented out. They need be made available, with the correct path
to the local JDK7 installation be specified in the 'jdk7home' property.

A Gradle installation is not required to run the build script.
The "Gradle wrapper" is already included in this commit to automatically
download the relevant Gradle software to use, with the "gradlew"
command.

The following commands should be executed in the Gitblit folder where
the build.gradle script resides, to build Gitblit software.

To produce all of the Gitblit artefacts, use the buildAll task:

  gradlew buildAll

The above command will generate all Gitblit software artefacts in
build/distributions folder, which include the Gitblit GO, WAR, Manager,
Federation Client, API client, and the site docs.

To build just the Gitblit GO software:

  gradlew buildGO

and the resulting gitblit-<version>.zip (for Windows) and
gitbit-<version>.tar.gz (for Linux) are produced in build/distributions
folder.

Similarly, to build just the site documentation:

  gradlew buildSite

and the Gitblit html documents are produced in build/site folder.

For the IDE tool such as Eclipse or IDEA development tools, the project
can be imported directly into the IDE, as they should understand and act
on the Gradle build script in the project. For example, in Eclipse, the
Gitblit project can now be imported directly via the File menu:

  Import --> Gradle --> Existing Gradle Project

And thus, all current files in the repository relating to Eclipse and/or
IDEA tools are deleted by this commit, and are added to the
.gitignore file, as they are now just for local development use.