Fix: Make CPU hog fix Java 7 compatible
The last fix for the stored config merged from Curly060 used Java8-isms.
In order to be able to include this fix in the next release, which will
be for 1.9, I have converted this to be compatible with Java 7.

Also, a file header was added to place it under APL.