ci: Allow to force build a nightly version
Enable a switch for manual workflow triggering that overrides the check
if a new commits were made since the last build. If the input is set to
'true', a new build is run even when no new commits are present.

This is mainly so that changes to the gitblit-docker repo can be included
in builds. This could also be automated in the workflow, if need be.